refac: Moved domain objects to domain (#317)
