+++ source/geh_calculated_measurements/src/geh_calculated_measurements/electrical_heating/domain/__init__.py
@@ -1,3 +1,12 @@
+from geh_calculated_measurements.electrical_heating.domain.calculated_names import CalculatedNames
+from geh_calculated_measurements.electrical_heating.domain.model.child_metering_points import (
+    ChildMeteringPoints,
+    child_metering_points_v1,
+)
+from geh_calculated_measurements.electrical_heating.domain.model.consumption_metering_point_periods import (
+    ConsumptionMeteringPointPeriods,
+    consumption_metering_point_periods_v1,
+)
 from geh_calculated_measurements.electrical_heating.domain.model.time_series_points import (
     TimeSeriesPoints,
     time_series_points_v1,
@@ -5,4 +14,13 @@ from geh_calculated_measurements.electrical_heating.domain.model.time_series_poi
 
 from .calculation import execute
 
-__all__ = ["execute", "TimeSeriesPoints", "time_series_points_v1"]
+__all__ = [
+    "execute",
+    "CalculatedNames",
+    "TimeSeriesPoints",
+    "time_series_points_v1",
+    "ChildMeteringPoints",
+    "child_metering_points_v1",
+    "ConsumptionMeteringPointPeriods",
+    "consumption_metering_point_periods_v1",
+]
