+++ source/geh_calculated_measurements/src/geh_calculated_measurements/electrical_heating/domain/calculation.py
@@ -8,10 +8,10 @@ from geh_common.telemetry import use_span
 
 import geh_calculated_measurements.electrical_heating.domain.transformations as T
 from geh_calculated_measurements.common.domain import CalculatedMeasurements, calculated_measurements_factory
-from geh_calculated_measurements.electrical_heating.domain import TimeSeriesPoints
-from geh_calculated_measurements.electrical_heating.infrastructure import (
+from geh_calculated_measurements.electrical_heating.domain import (
     ChildMeteringPoints,
     ConsumptionMeteringPointPeriods,
+    TimeSeriesPoints,
 )
