+++ source/geh_calculated_measurements/src/geh_calculated_measurements/electrical_heating/domain/transformations/metering_points.py
@@ -7,10 +7,8 @@ from pyspark.sql import DataFrame
 from pyspark.sql import functions as F
 
 from geh_calculated_measurements.common.domain import ColumnNames
-from geh_calculated_measurements.electrical_heating.domain.calculated_names import (
+from geh_calculated_measurements.electrical_heating.domain import (
     CalculatedNames,
-)
-from geh_calculated_measurements.electrical_heating.infrastructure import (
     ChildMeteringPoints,
     ConsumptionMeteringPointPeriods,
 )
