+++ source/geh_calculated_measurements/src/geh_calculated_measurements/electrical_heating/infrastructure/__init__.py
@@ -1,5 +1,3 @@
-from .electricity_market.child_metering_points.wrapper import ChildMeteringPoints
-from .electricity_market.consumption_metering_point_periods.wrapper import ConsumptionMeteringPointPeriods
 from .electricity_market.repository import Repository as ElectricityMarketRepository
 from .measurements.measurements_gold.database_definitions import MeasurementsGoldDatabaseDefinition
 from .measurements.repository import Repository as MeasurementsGoldRepository
@@ -7,8 +5,6 @@ from .measurements.repository import Repository as MeasurementsGoldRepository
 __all__ = [
     # Electricity market repository, types, and database definitions
     "ElectricityMarketRepository",
-    "ChildMeteringPoints",
-    "ConsumptionMeteringPointPeriods",
     # Measurements core repository, types, and database definitions
     "MeasurementsGoldRepository",
     "MeasurementsGoldDatabaseDefinition",
