+++ source/geh_calculated_measurements/src/geh_calculated_measurements/electrical_heating/infrastructure/electricity_market/repository.py
@@ -2,17 +2,11 @@ from pyspark.sql import DataFrame, SparkSession
 from pyspark.sql import functions as F
 from pyspark.sql import types as T
 
-from geh_calculated_measurements.electrical_heating.infrastructure.electricity_market.child_metering_points.schema import (
-    child_metering_points_v1,
-)
-from geh_calculated_measurements.electrical_heating.infrastructure.electricity_market.child_metering_points.wrapper import (
+from geh_calculated_measurements.electrical_heating.domain import (
     ChildMeteringPoints,
-)
-from geh_calculated_measurements.electrical_heating.infrastructure.electricity_market.consumption_metering_point_periods.schema import (
-    consumption_metering_point_periods_v1,
-)
-from geh_calculated_measurements.electrical_heating.infrastructure.electricity_market.consumption_metering_point_periods.wrapper import (
     ConsumptionMeteringPointPeriods,
+    child_metering_points_v1,
+    consumption_metering_point_periods_v1,
 )
